Fix Django 2.2 in tests
Happily this does seem to have been a test specific issue.

It's not clear why this started being an issue in 2.2 as the order
in which the tables are created is the same. I suspect that it may
be related to the changes to the test runner around multi-db testing
and/or the database backend changes. In any case, forcing all apps
to load their tables from models rather than migrations fixes this.